#7 Server component response header order.
#7 Earlier Server component has generated the response header creating a
new SOAP message with an empty header, then adding required header
elements and copying all the values from the request message header.
Header values have remained the same, but the order of the elements
might have changed. The Server component has hard coded order for header
elements when a new response message is generated. According to X-Road
message protocol the service MUST copy all the header fields from the
request to the response in the exact same sequence with the exact same
values.

Now the Server component clones the request message, removes all the
elements from the body and uses the cloned request message as a template
for the response message. In this way the order and the values of the
SOAP header are kept.